# swagger/v1/swagger.yaml
swagger: '2.0'
info:
  description: >-
    SMX api description.
  version: 1.0.0
  title: MoneyApi
  termsOfService: ''
  contact:
    email: 
host: 107.22.146.137
basePath: /api/v1
...
schemes:
  - http
securityDefinitions:
  X-API-APIKEY:
    name: X-API-APIKEY
    in: header
    description: logged in user api key
    type: apiKey
paths:
...
  /countries:
    get:
      tags:
        - country
      summary: Get the list of countries
      description: The endpoint will get the list of countries in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      responses:
        '200':
          description: List of countries
          schema:
            $ref: '#/definitions/Countries'
        '401':
          $ref: '#/responses/401'
        '500':
          $ref: '#/responses/500'
  /states:
    get:
      tags:
        - country
      summary: Get the list of states
      description: The endpoint will get the list of states in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      parameters:
        - in: query
          name: country[c]
          description: Alpha3 code of country
          type: string
          example: "usa"
          required: true
      responses:
        '200':
          description: List of states
          schema:
            $ref: '#/definitions/States'
        '401':
          $ref: '#/responses/401'
        '500':
          $ref: '#/responses/500'
  /countries/countries_info:
    get:
      tags:
        - country
      summary: Get the list of countries with info
      description: The endpoint will get the list of countries with info in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      responses:
        '200':
          description: List of countries with info
          schema:
            $ref: '#/definitions/CountriesInfo'
        '401':
          $ref: '#/responses/401'
        '500':
          $ref: '#/responses/500'
...
  /countries/smx_countries:
    get:
      tags:
        - country
      summary: Get the list of SMX countries
      description: The endpoint will get the detail list of SMX countries in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      responses:
        '200':
          description: List of SMX countries
          schema:
            $ref: '#/definitions/Countries'
        '401':
          $ref: '#/responses/401'
        '500':
          $ref: '#/responses/500'
  /countries/flag:
    get:
      tags:
        - country
      summary: Get the country flag
      description: The endpoint will get the country flag in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      parameters:
        - in: query
          name: name
          description: Name of the country
          required: true
          type: string
      responses:
        '200':
          description: Country flag
          schema:
            $ref: '#/definitions/CountryFlagResponse'
        '401':
          $ref: '#/responses/401'
        '404':
          $ref: '#/responses/404'
        '500':
          $ref: '#/responses/500'
  /countries/update_flag:
    patch:
      tags:
        - country
      summary: Update country flag
      description: The endpoint will update flag of country in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      parameters:
        - in: body
          name: body
          schema:
            properties:
              country:
                properties:
                  name:
                    type: string
                    description: Country name
                    example: "Zimbabwe"
                  flag:
                    type: string
                    description: Country flag
      responses:
        '200':
          description: Successful Operation
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          $ref: '#/responses/400'
        '500':
          $ref: '#/responses/500'
...
definitions:
...
  CountryResponse:
    allOf:
      - type: object
        properties:
          name:
            type: string
            description: Country name
            example: "Benin"
          alpha_code:
            type: string
            description: Alpha code
            example: "BEN"
          country_code:
            type: string
            description: Country code
            example: "1"
          phone_number_lengths:
            type: integer
            description: Phone number length
            example: 8
          flag:
            type: string
            description: Country flag
  Countries:
    properties:
      countries:
        type: array
        items:
          $ref: '#/definitions/CountryResponse'
  CountriesInfo:
    properties:
      countries:
        type: array
        items:
          allOf:
            - type: object
              properties:
                name:
                  type: string
                  description: Country name
                  example: "Benin"
                alpha_code:
                  type: string
                  description: Alpha code
                  example: "BEN"
                country_code:
                  type: string
                  description: Country code
                  example: "1"
                phone_number_lengths:
                  type: integer
                  description: Phone number length
                  example: 8
  CountryFlagResponse:
    allOf:
      - type: object
        properties:
          updated_at:
            type: string
            description: Country flag update datetime
            format: date-time
            example: '2021-07-15T18:08:36.189Z'
          flag:
            type: string
            description: Country flag
  States:
    properties:
      states:
        type: array
        items:
          allOf:
            - type: object
              properties:
                name:
                  type: string
                  description: State name
                  example: "Alaska"
...
  Error:
    type: object
    required:
      - errors
    properties:
      errors:
        type: array
        items:
          type: string
          description: error message
  SuccessResponse:
    type: object
    required:
      - status
    properties:
      status:
        type: string
        description: Sucecess message
...
responses:
  '400':
    description: Input invalid or Bad Request
    schema:
      $ref: '#/definitions/Error'
  '401':
    description: Token Invlaid
    schema:
      $ref: '#/definitions/Error'
  '404':
    description: Resource not found
    schema:
      $ref: '#/definitions/Error'
  '500':
    description: Internal Server error
    schema:
      $ref: '#/definitions/Error'